Update tiles, observation, and count with map area query params
