+++ src/components/Explore/ExploreContainer.js
@@ -1,6 +1,7 @@
 // @flow
 
 import { useRoute } from "@react-navigation/native";
+import { searchObservations } from "api/observations";
 import {
   EXPLORE_ACTION,
   ExploreProvider,
@@ -12,8 +13,8 @@ import {
   WILD_STATUS
 } from "providers/ExploreContext.tsx";
 import type { Node } from "react";
-import React, { useEffect, useState } from "react";
-import { useCurrentUser, useIsConnected } from "sharedHooks";
+import React, { useCallback, useEffect, useState } from "react";
+import { useAuthenticatedQuery, useCurrentUser, useIsConnected } from "sharedHooks";
 
 import Explore from "./Explore";
 
@@ -148,6 +149,13 @@ const ExploreContainerWithContext = ( ): Node => {
 
   const currentUser = useCurrentUser();
 
+  const [count, setCount] = useState( {
+    observations: null,
+    species: null,
+    observers: null,
+    identifiers: null
+  } );
+
   const { state, dispatch, makeSnapshot } = useExplore();
 
   const [region, setRegion] = useState( {
@@ -221,25 +229,63 @@ const ExploreContainerWithContext = ( ): Node => {
     } );
   };
 
+  const updateCount = useCallback( newCount => {
+    setCount( {
+      ...count,
+      ...newCount
+    } );
+  }, [count] );
+
   const filteredParams = mapParamsToAPI(
     state,
     currentUser
   );
 
+  const queryParams = {
+    ...filteredParams,
+    per_page: 20
+  };
+  delete queryParams.taxon_name;
+
+  const paramsTotalResults = {
+    ...filteredParams,
+    per_page: 0
+  };
+
+  // 011224 amanda - we might eventually want to fetch this from useInfiniteObservationsScroll
+  // instead of making a separate query, but per_page = 0 should make this extra query a low
+  // performance cost
+  const { data } = useAuthenticatedQuery(
+    ["searchObservations", paramsTotalResults],
+    optsWithAuth => searchObservations( paramsTotalResults, optsWithAuth )
+  );
+
+  useEffect( ( ) => {
+    if ( data?.total_results && count.observations !== data?.total_results ) {
+      updateCount( { observations: data?.total_results } );
+    }
+  }, [data?.total_results, updateCount, count] );
+
+  const closeFiltersModal = ( ) => setShowFiltersModal( false );
+
+  const openFiltersModal = ( ) => {
+    setShowFiltersModal( true );
+    makeSnapshot( );
+  };
+
   return (
     <Explore
-      exploreAPIParams={filteredParams}
-      region={region}
-      exploreView={exploreView}
       changeExploreView={changeExploreView}
-      updateTaxon={updateTaxon}
+      closeFiltersModal={closeFiltersModal}
+      count={count}
+      exploreView={exploreView}
       isOnline={isOnline}
+      openFiltersModal={openFiltersModal}
+      queryParams={queryParams}
+      region={region}
       showFiltersModal={showFiltersModal}
-      openFiltersModal={() => {
-        setShowFiltersModal( true );
-        makeSnapshot( );
-      }}
-      closeFiltersModal={() => setShowFiltersModal( false )}
+      updateCount={updateCount}
+      updateTaxon={updateTaxon}
     />
   );
 };
